move types and kinds into separate modules
