+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lsts"
-version = "0.2.4"
+version = "0.3.0"
 authors = ["Andrew <andrew@subarctic.org>"]
 license = "MIT"
 description = "Large Scale Type Systems"
