+++ src/lib.rs
@@ -4,3 +4,4 @@ extern crate pest_derive;
 
 pub mod tlc;
 pub mod typ;
+pub mod kind;
